[tagged] Port Code to HeapObjectLayout
Ports the Code object. Notable: DeoptimizableCodeIterator::Next
returned Code() which under a V8_OBJECT class constructs a real object
on the stack and wraps its address, causing callers looping on
!x.is_null() to spin on a bogus heap pointer. Fixed to return {}.
ExternalPointerMember fields in serializer paths use ExternalPointerSlot
built directly from the member's storage address.

Bug: 42202654
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/7789937
Commit-Queue: Jakob Linke <jgruber@chromium.org>
Cr-Commit-Position: refs/heads/main@{#106816}

+++ src/compiler/backend/x64/instruction-selector-x64.cc
@@ -137,7 +137,7 @@ bool CanBeImmediate(InstructionSelector* selector, OpIndex node) {
         const int64_t value = constant.smi().ptr();
         return ValueFitsIntoImmediate(value);
       } else {
-        const int32_t value = constant.smi().ptr();
+        const int32_t value = static_cast<int32_t>(constant.smi().ptr());
         // int32_t min will overflow if displacement mode is
         // kNegativeDisplacement.
         return value != std::numeric_limits<int32_t>::min();
